Rename personal details controller to base controoller
Move the controller into a personal_details folder.
